requirements.txt
PyYAML==6.0.1
selenium==3.141.0
urllib3==1.26.18
robotframework==6.0
pylint==3.1.0
robotframework-pabot==2.18.0
robotframework-tidy==4.11.0
jira==3.8.0
robotframework-datadriver==1.11.1
wheel==0.43.0
robotframework-lint==1.1
#eyes_robotframework==5.4.0
openpyxl==3.1.2
requests==2.29.0
robotframework-seleniumlibrary==5.1.3
lxml==5.1.0
beautifulsoup4==4.12.3
robotframework-imaplibrary2==0.4.6
pandas==2.0.3
numpy==1.24.4
easyocr==1.7.2
pytesseract==0.3.13
filelock==3.16.1